requirements.txt
Django
dj-database-url==0.2.2
django-dotenv==1.2
django-getenv==1.2
raven==4.0.4
wsgiref==0.1.2
newrelic